Update to MaterialModel class
The MaterialModel is now a functional elastic material model.  While
the interfaces may still need some adjustments, the class can be used
as a base for developing other material models.

I've added a test that tracks a stress in xx that is rotated 90
degrees and ends up as a stress in yy.

r2733