Fix bug in TransientExecutioner
If a solve failed, the time would be set to the old time plus the
failed delta t plus the new delta t.  With that fixed, the gap heat
transfer analysis runs to completion with no problems.

Also reworked the distance calculation of the gap heat transfer code
to account for roughness.

r2913