PresetBC
Added a new PresetNodalBC class that is a NodalBC (Dirichlet) with a
minor twist.  It puts its value into the solution vector at the
beginning of the solve.  This causes the residual at the given node
and for the given variable to be zero from the beginning of the
nonlinear solve.

r4667