Added new input file for natural convection in the problems/bighorn area.
Added NodalMomentumInviscidFlux kernel.

This kernel uses nodal aux vars *only*
for computing its values.  Also working on the Jacobian and
OffDiagJacobian values for this kernel.

Changes to NSMomentumInviscidFluxAux:

The NSMomentumInviscidFluxAux auxkernel is coupled to all momentums
and all velocities, the flux_vector_subscript and equation_index
determine which component of which flux vector is computed.

Compute the on-diagonal Jacobian value for NodalMomentumInviscidFlux.

Added x and y aux kernels for y-equation.

Use GlobalParams section in input file.

Change all GlobalParams to RequiredParams in classes,
avoid setting default values for a given parameter in
multiple places.

r5016