Common base class for solid mechanics material models
The linear models (those derived from SolidMechanicsMaterial) were not
touched.  Because these may be used for adaptivity, and because
adaptivity and state variables are incompatible, these models could
not be converted.

This change consolidates a significant portion of code, including the
code controlling smeared cracking.  (Smeared cracking is not available
in the linear models.)

r5832