Radial power factor through a function
Implemented radial power factor using BurnupFunction.

BurnupFunction uses the approach in Lassman (Jou. Nuc. Mat., 208
(1994), p 223-231).  The function computes fission rate, burnup,
radial power factor (which is applied to the fission_rate
automatically), and the concentrations of 235, 238, 239, 240, 241, and
242.  Fission rate and burnup need to be transferred to the finite
element mesh.  The other quantities may be if needed in the output
file.

This capability is intended to be used with the Burnup block, which is
tied to three Actions.  Using this block will register the auxiliary
variables, create the auxiliary kernels, and create the
BurnupFunction.  This will cause the fission rate and burnup to be
transferred to the finite element mesh automatically.

It is possible to turn off the radial power factor and use simply the
rod-averaged linear power and an axial power profile.  In that case,
the concentrations and radial power factor will not be computed.

Ref #000

r6518