Misc. fixes to workshop slides:
In ex08, call the boundaries top and bottom instead of left and right in the input file.

Put actual types in Dirac function prototypes slide.

Fixed outline slides in C++ section, and simplifed intrinsic data types slide.

More spacing and formatting fixes in the C++ section of the workshop.

r6572