Add optional loadstep predictor
the command "predictor_scale = " in the transient executioner can now
be used to enable a predictor that extrapolates previous solutions to
get a better starting point for the next solution.  The scale factor
can range from 0 to 1, with 1 doing the full extrapolation and 0
doing none.

The initial residual is computed prior to applying the predictor or
preset bcs, so that convergence is relatative to the same initial
residual whether or not those features are used.

Addresses #712

r6936