Allow piecewise linear functions to be functions of axis
Add the optional "axis = [0,1,2]" command line to the PiecewiseLinear
function.  If that line is absent, the default behavior of being a
function of time is retained.  If the axis is specified, it is a
function of the position on that axis. Closes #756

Use this option in a new test of relocation driven by temporally and
spatially varying q defined with 2 piecewise linear functions used
together with a composite function.

r7462