Generate error if thermal contact order differs from order of temperature variable
Previously, the code would segfault if you used a second order
variable for temperature and didn't specify an order for the
ThermalContactAction because it defaults to FIRST.  I made the code
check for this and generate an error if that happens.

A better solution would be to have the ThermalContactAction determine
what order the temperature variable is and have it create all aux
vars and bcs with that order.  That doesn't appear to be possible
because the variables don't exist when ThermalContactAction::act() is
called.

Closes #761

r7504