Generate error if wrong type is used for input parameter
If you accidentally provided the wrong type in an input line, the code
would silently set that parameter to whatever the default value for the
expected type was.  For example, if you had the line:
axis=z
in your input file, you would get axis=0, because it was expecting an int.

This commit makes that generate an error now, and fixes a few tests that
were using the wrong type for input parameters. Closes #757

r7738