Toward more unified solid mechanics materials
A step toward eliminating the RZ variants of the material models.

Created a set of Element classes that incorporate the element
formulation.  These classes will primarily compute the mechanical
strain.  The computation of the thermal strain, volumetric strain,
cracking, and stress are left to the SolidModels (Materials).

With this checkin, the Element classes compute the strain.  In a
subsequent checkin, I will unify the material models.

r7875