Add option to PiecewiseLinearFile function to provide data in columns
The PiecewiseLinearFile function previously took only comma-separated
data arranged in two rows: one for the x data and one for the y data.
There is a new optional parameter:
format = rows|columns
to control whether the data is formatted in rows or columns.  Rows
is still the default, and this format parameter is not required.

Also allowed for comments in the data file and for the data to be
separated by any combination of whitespace and commas.

I worked on this with Steve Novascone.  Addresses #825

r8246