Add HeatSource kernel
This is a basic volumetric heat source kernel that doesn't rely
on fission rate like the neutron heat source does.  It doesn't do
much -- just derives from BodyForce.  Difference is that it has a
name that makes sense for thermal analysis, and "value" defaults
to 1 rather than 0, so you don't have to supply a value if using
a function.  Also added a verification test.

r10771