Another fix to divergence check in nonlinear convergence check closes #1490
There was another error in the divergence check.

Also, the divergence check was based on the initial residual, which is calculated
before preset BCs modify the solution, so it's potentially orders of magnitude
smaller than the first nonlinear residual.  Change the divergence check to be based
on the initial nonlinear residual instead.

r14620