Add ReferenceResidualProblem in elk
This allows the user to optionally provide a postprocessor or set of
postprocessors to use as a reference quantity to be used in place of
the initial residual in the nonlinear convergence check.

If you use the nodal l2 norm of the saved residual as this criterion,
the solver will converge relative to a reasonable measure of the
loading on the system, rather than the initial residual, which has
little physical meaning for the model.  Doing this permits the solver
to take minimal iterations when the load is not changing.

The regression test included in this commit demonstrates the usage of
this option.

r15295