UO Refactor - closes #1622
The hierarchy has been changed so that variables are no longer included by default.
Developers should use the normal coupling interface to consume variables in both
UOs and PPs.  Additionally, there are multiple versions of "Integral" classes,
one that takes a variable and one that does not for each major type.

r15901