Add option to smooth normals for contact near face boundaries
You can add the parameter:
normal_smoothing_distance = <value>
to cotact blocks to smooth the normal as the node transitions from one
face to another.  The specified distance is in parametric coordinates.
Smoothing starts to take place when the contact point is within that
distance from an edge.

The default value of this parameter is 0, which means that no smoothing
takes place, so the default behavior is unchanged.  It is recommended
to set this parameter to something around 0.1.

This commit adds tests for this in 2D.  It still needs testing for 3D.

Ref #1874

r18034