Add option for contact normal smoothing based on nodal normals Ref #1874
The nodal surface normals can now optionally be used as a basis for
contact normal smoothing.  The nodal normals are simply interpolated
to the contact point.

This option can now be optionally enabled within all contact or similar blocks that use the PenetrationLocator by using the command:
normal_smoothing_method = nodal_normal_based
The default behavior is still to use contact normals based on smoothing
between the normals for the faces that share a common edge (edge_based option).

If the nodal_normal_based option is used, this block must be present to enable calculation of nodal normals:
[NodalNormals]
[]
Note that the corner_boundary option must be used in NodalNormals to prevent smoothing around corners at the edges of the contact surface.

r19533