Fix issue with contact smoothing when on a ridge between two faces
Ref #1874

If the closest contact point is on a ridge between two faces, we
were not changing the reference coordinates of that point to
the point on the face.  This made the smoothing based on nodal
normals behave incorrectly.

Also, delete an unused method in PenetrationLocator and fixed an
issue with handling nodes off of edges of tri faces.

r19561