Normal smoothing cleanup Ref #1874
Consolidate duplicated code to look up sides for an element in the master boundary
and create a PenetrationInfo object for that side.
Correctly handle cases where multiple sides on an element belong to master boundary.
Skip creation of PenetrationInfo objects in some cases where a node isn't on the face.
Fix some shadow variables.

r19885