Picking a better name for _is_parallel (it really means _is_nemesis).
The accessor is removed, but it could always be readded.

'nemesis' is a parameter on MooseMesh instead of FileMesh.

Reformatting long doc strings onto separate lines.

Adding new command line option to MOOSE: --parallel-mesh.

MooseMesh now uses the MooseApp object to determine if
--parallel-mesh has been specified on the command line.

MooseMesh reads an enumeration for the type of mesh "distribution" to use.

Note: if the user does not specify anything for 'distribution' in the
Mesh block, it is possible to set it to parallel via the
--parallel-mesh command line argument, otherwise it defaults to
serial.  This command line arg is not allowed to override the user if
he does specify 'distribution = serial' in the Mesh block.

Refs #2105.

r20832