More Transient timestepping cleanup Ref #2265
If a sync time was hit, Transient was ignoring constraints on the timestep
the next step.  I removed that, and it only affected one test, which had
a sync time that wasn't being honored because of this.

Delete unused variables _unconstrained_dt, _unconstrained_dt_old.  Rename _reset_dt
in Transient to _force_output to be more descriptive of what it does.  Not to be
confused with _reset_dt in the TimeStepper.

r22531