Optimizing couling matrix entries (closes #2535)
For computing "full" matrix Jacobian, we store the list of blocks we
need to visit. Each block has a pair of unsigned ints identifying its
position the global matrix. When we go over the list, we take the number
and call `SystemBase::getVariable`. In this call we downcast to
MooseVariable and run checks agains NULL.  This is not necessary, since
we can just store `MooseVariable *` in the coupling list and use those
directly.

Motivation behind this: I saw about 2% of runtime being spend in
`Assembly::prepare`. After this patch the time looks like 0.2%.