Adding ability to append to Tecplot ASCII files.
Adding a test for this capability.

This feature basically consists of writing multiple Tecplot ASCII
files into the same file using std::ofstream::app, which Tecplot is
indeed able to read.  The drawback is that the Mesh is written to the
file each time. I have a question in to Tecplot support to see if it
would be possible to have only one copy of the Mesh in the file,
similar to the way the Exodus file format works.

Appending is not currently supported for binary files, if you have
selected the binary output format for Tecplot, the ascii_append flag
will be ignored.  I'm also investigating appending to binary files, so
this interface may be updated in the future.

Refs #440.