Error checking for crack tip nodeset with multiple nodes in 2d closes #2963
For the case where the J-Integral is evaluated in 2d (even in a 3d model),
only only allow the crack tip nodeset to contain multiple nodes if they're
collinear.

Also, use range-checked parameter for 2d axis.  I had to change the name of
the parameter from 2d_axis to axis_2d to make the range checking parser happy.