Change the behavior of --mesh-only slightly.
When a filename is not provided, or when an Exodus file is requested,
we now write it using the "logical" dimension of the Mesh rather than
the spatial dimension (which is the default).  This helps us better
support visualizing 2D triangular meshes in peacock, since they will
now show the correct sidesets.

Interestingly, this did not require regolding any of the --mesh-only
MOOSE tests, since they were all already 3D.  But you can see the
issue if you open $MOOSE_DIR/test/tests/mesh/gmsh_bcs/plate_hole.msh
in peacock before and after the patch.

Refs #1568.