Add methods to determine whether node or element is on boundary ref #3112
This adds MooseMesh::isBoundaryElem(), which is equivalent to
MooseMesh::isBoundaryNode().  Add two variants of this method: one to
determine whether the element is adjacent to any boundary, and one to
determine whether the element is adjacent to a specified boundary.  Also
add a version of isBoundaryNode() in which the boundary is specified.