Fixing uninitialized _coord_type (closes #3231)
When using node face constraint and one side is using elements with
different dimension, _coord_type in Assembly::reinitNeighborAtReference
may be uninitialized. The actual problem was that the _coord_type was a
member variable set by previous calls to reinit(). Since all variables
inside reinitNeightborAtReference are local, also _coord_type has to be
local and it has to query the neighbor element to get the right value.