Shifted van-Genuchten curve added.  Fixes #3350
This is for 2-phase flow.  It makes the Jacobian nonsingular
in 2phase models when one phase disappears and the residual
saturation is zero.  In limited testing this appears to give
good convergence characteristics.

Included are doco and tests.

Also, I have changed RichardsSeff::dseff and d2seff so they
accept references to the answers, rather than providing the
answers as return values.  According to jwpeterson this might
speed up computations by not having to do extra copying.
Since I have so many RichardsSeff type of classes, this required
a lot of minor changes