Weak-plane shear plasticity added.  Fixes #3535
A number of tests have been added that demonstrate correctness
for small and large deformations.

Fairly comprehensive doco can be found on the mooseframework.org wiki.

Please don't be too harsh in judging the coding quality.  I have
great plans for these plasticity Material classes, and this present
class will probably change greatly over the next few weeks, but
since this works as planned i'd like it to be merged now, so that
people can play and comment on things.  Otherwise the changes in
"a few weeks" will be far too big for one PR.