This is one NASTY commit!
Here I add ReturnMapUtils, and CPPUNIT-style tests of it.  The
point of ReturnMapUtils is to solve the NASTY linear system
that is created during the Newton-Raphson process of the
plastic return-map algorithm.  I made a namespace so that i
could most easily test this using CPPUNIT separately to any
plasticity stuff

Part of #3573