Substantially refactored FiniteStrainPlasticBase by including most of ReturnMapUtils into it.
Also included a lot of debugging features that will be useful
for future developers to debug their plasticity models (eg,
debugging dyieldFunction_dstress).

Also reworked calculateConstraints, calculateRHS, etc, so that
the code is more clear (might be slightly less efficient this
way, but that's too bad for now).

Altered tests so they crash if max_iterations is exceeded or
the line search fails.