Model type parameter lives in FlowModel (closes idaholab/relap-7#13)
Each component that has the flow equations on it inherits form
FlowModel, which hold _model_type variable that is set from an input
parameter. It it is not specified loally, it is taken from the global
params. This hold for Pipes and boundaries.  We, however, used FlowModel
for junctions, which is not a good idea, because they might connect
pipes with different models.