Hardening added to WeakPlaneTensile.
Three nice tests added.

Also, FiniteStrainPlasticBase now provides a Material which is
the number of Newton-Raphson iterations used.  This is useful for
debugging, but also i think it'll be useful in the future when
we have to apply the strain increment in small amounts in order
to return to the yield surface for very nonlinear models.

Fixes #3625