Made eigenvalue and doublecontraction methods in RankTwoTensor const
Also eigenvalue routine "syev" in RankTwoTensor now explicitly
symmeterises the Tensor instead of using its upper-diagonal only.
This is convenient in finite-differencing derivatives.

Also, i try to be clever with the equal-eigenvalue case in
dsymmetricEigenvalues in RankTwoTensor where there is a discontinuity
in the derivative.  I now use a form that agrees with central differences.
This is tested in CPPUNIT.  Hopefully this will not be activated too
often in real plasticity models as discontinuous derivatives is
not a good thing in Newton-Raphson!

Added sin3Lode and derivatives to RankTwoTensor.  Tested in CPPUNIT