Add CrackFrontData PostProcessor ref #3671
This postprocessor currently calculates the value of a user-specified variable at
a specified node index along the crack front used for J-Integral calculations.

I'll expand it to output other data about the crack front as well.

Next step is to add an option to JIntegralAction to create one of these for
each point along the crack.