Rename JIntegralAction to DomainIntegralAction, make general ref #3680
The JIntegral block in the input file is replaced with DomainIntegral,
and an additional parameter "integrals" must be specified, which contains
a list of integrals that are to be computed.  This paves the way for adding
the interaction integrals, and having the option to compute multiple types
of domain integrals in one run.