Add option to action to output variable values along crack ref #3671
This adds the parameter 'output_variable', which the user can specify
to create PostProcessors that output the values of variables along
the crack.  The naming scheme of those PostProcessors matches that of
those used to calculate domain integrals.  This will allow comparing
toughness to stress intensity at locations along the crack.