Make J Integral tests more tolerant of machine differences ref #3633
For these tests, the nonlinear residual won't drop any lower than
about 5e-4, which translates to a relative tolerance of about 1e-8.
The domain integrals are pretty sensitive to small differences in
the solution, so I've had to use pretty loose tolerances to get
this test to pass cross-platform, and it's still been problematic.

I've tried a lot of things to try to get this to converge tighter,
and haven't found anything that makes a difference.

That convergence floor seems to be independent of the load, so I
increased the loading by a factor of 1000, which permits solving
to a much tighter relative tolerance.  I also made these use the
default setting for pc_hypre_boomeramg_strong_threshold.