Nonzero initial stress components may be specified as functions in TensorMechanicsMaterial
Fixes #3763

I initially wanted to specify the components as AuxVariables, but
i couldn't get it to work because the AuxVars weren't defined at
the point of calling initQpStatefulProperties.  Functions will do
for me for the present, but if anyone better at moose would like
to change to AuxVariables, i'd be quite happy!