Fixing du_dot_du for non-nodal FEs (closes #3784)
The computation of du_dot_du assumed nodal FE and was putting the
computed value into each quad-point. This did not work for non-nodal FEs
like monomials. Those require the "function" to be projected into the
space.  Fortunately, all of our time scheme have du_dot_du as a scalar
value, thus, we do not need the solution vector and we can store it as a
single value.

Also, adding a test case that will check the Jacobian by FDing it (on
monomial FEs).