Add ability to transfer postprocessors from MultiApps ref #4028
This commit adds the ability to use MultiAppPostProcessorTransfer to
transfer postprocessor values from subapps to a postprocessor in the
master app.  To do this, you have to specify a value for the new
'reduction_type' parameter, which can have values of 'average', 'sum',
'minimum', and 'maximum'.  This commit also adds a test for this
capability.