Add InitialConditions to the _all_objects vector.
This way, they are returned correctly by Warehouse::all().  I don't
actually know if this is useful, though, since there are several
different types of ICs and they can be block-restricted and inactive.
Also, ScalarInitialiConditions are not InitialConditions, so they
can't be stored in the _all_objects vector of the
InitialConditionWarehouse.

Refs #3692