Fixing unit test for eigen() and eigenNonsym()
These calls take a ColumnMajorMatrix for storing a vector of eigen
values. Then, these matrices are resized into vectors (one column, n
rows, where 'n' is the number of eigen values).  On the caller side, the
tests actually assumes a matrix and is trying to access its values that
are out of range.

Thus, not passing a matrix, but vector and not accessing the OOB
indices.

Closes #4516