Option to define crack front with points for fracture domain integral ref #4936
This permits the user to define the crack front using a set of points
defined with the command 'crack_front_points' instead of 'boundary'.

Almost everything works with this option.  The following can't yet be used
(and errors will be generated if they are):
t-stress
definining intersecting surfaces with sidesets
extracting solution variables on crack front