Fix Dirac point caching in parallel
As pointed out by @WilkAndy, the Dirac point caching stuff was not
working quite right in parallel.  This was because processors only
cache local (Point, Elem) information, and were still calling the
PointLocator to try and find Points which were guaranteed not to be
there.  The fix is to return NULL on those processors, which is what
they were going to return anyway.

Refs #739.