Updating cracking_exponential.i.
Switch from end_time to num_steps to avoid a weird corner case of
needing to take an O(1.e-13) timestep (due to roundoff error over 1000
timesteps) to hit the final time.

This test has so many timesteps, it is necessary to just re-gold it
rather than try and get it to match the old gold file, which sometimes
converged on the old absolute tolerance of 1.e-8, and sometimes converged
on the old relative tolerance of 1.e-4.

Refs #5066