Updating gap_heat_transfer_htonly_rspherical.i
I don't really understand what this test does.  It has a super-loose
nl_abs_tol = 5e-2, and does not actually perform any Newton steps
after timestep 10.  This may be intentional, so I just left it as-is.
The new, tighter nl_rel_tol that I added also required regolding the
test, but only for a slight diff in the second timestep.

Attn: @jasondhales

Refs #5066