Updating combined/contact tests to use relative tolerances.
* 4ElemTensionRelease.i
* 4ElemTensionRelease_mechanical_constraint.i
* 8ElemTensionRelease.i

The use of a relative tolerance is complicated in these examples by
the existence of a Predictor, which in the later timesteps causes the
initial residual seen by PETSc to be very small.  Therefore we have
also included a (fairly small) nl_abs_tol in this test.  You can
remove the nl_abs_tol if you also remove the Predictor section of the
input file.

This example also demonstrates an issue with MOOSE's previous default
value of 'compute_initial_residual_before_preset_bcs = true' in the
interaction between PresetBCs and Predictors: the residual computed
before applying the PresetBCs was exactly 0.00, and this meant the
nl_rel_tol could never be satisfied.  This issue becomes irrelevant
with the new default value. Since the initial residual computed by
PETSc is correct, we are able to satisfy the requested relative
tolerance.

Refs #5066