Updating the KKS example in the phase_field module for PETSc-3.6
PETSc-3.6 changed the default value of -pc_factor_shift_type to
'none'.  This affected the KKS example, since for some reason it
produces a finite-difference Jacobian with all zero entries... Also
make this example fail if it tries to cut the timestep rather than
taking a bunch of failed timesteps.

Refs #5162