Fix glued contact capture closes #5405
Capture nodes in contact if the node is right on the surface, rather
than requiring it to be penetrated.  Also, add an optional parameter,
capture_tolerance, which can be used to specify a distance that a node
can be off the surface and still be captured.  This is useful for
glued contact of initially curved surfaces where there might be a small
gap due to numerical roundoff.