Minor and rather trivial cleanup of tests.
The Peacock bug that Heather found seems to be activated by an expression like
   E = 0.3  # Young's = 234
It appears to need a valid input (E=0.3) followed by a hash, followed by a single quote followed by an equals sign.

I removed all such things from the TensorMechanics and Richards tests, and took the opportunity to clean up some doco in some tests, and remove a number of unnecessary things from quite a few tests.

I'm not sure if this really closes this issue.  Possibly someone wants to fix Peacock so that it correctly parses Heather's problematic line.

#5270