Fix incorrect assumption in FEProblem::computePostCheck().
I initially thought that the search direction, y, from PETSc was
simply y=w-x but on further reflection if that were the case, why
would they bother passing all three vectors to you?  This commit fixes
the test in BISON which was failing because it used Dampers.  We have
tests in MOOSE that use Dampers, but they didn't seem to be affected
by this bug.