Recompute Jacobian if dt changes in an explicit scheme
We need to recompute the Jacobian if dt changes. We assumed constant
dt for explicit solves so we set _const_jacobian flag to true. This can
be done only if dt_old and dt are the same. It saves Jacobian
evaluation, which is what we want in an explicit scheme.

Closes #5591