This adds side-by-side tests of the TensorMechanicsPlasticJ2 user object and the Solid Mechanics LinearStrainHardening material.
The TM code is taking appoximately 7x as long as SM code to run the same problem. These files will be used for profiling in attempt to figure out where the TM code gets stuck.

There is also a folder called 'necking' with simple 2D necking tests, which illustrate the same pattern for run time.

Refs. #5628.