Fix for issue #3525
Remove deprecated moab adapters.  The removed code relied on the fact that DTK had a required dependence on moab.  This dependency has been removed for DTK2 and the moab adapters no longer compile with updated DTK2 library. This comit removes the deprecated code so that MOOSE can build against DTK2.