Add 'noproxy' flag to the ICEUpdater.
This allows us to control whether proxies are ignored while making
libcurl requests.  The flag has to be passed from ICEUpdater ->
Updater -> UpdaterThread -> LibcurlUtils where it is actually used,
which is a bit awkward.  If we ever need to pass any more flags, we'll
have to jump through similar hoops.

Refs #5663.