Removes ControlInterface, which is not needed.
ControlInterface was added to allow Controls and the reporter postprocessor to share get methods, but
since the reporter only needs to read a parameter it doesn't need full access. Thus, the interface was removed
making the Control class complete and remove then chance of people trying to inherit from ControlInterface.

(refs #5680)