Delete invalid MultiMooseEnum tests.
These tests never ran in opt mode because they segfault.  So, they
were guarded to run only in debug mode, but that doesn't work either
because they assert before an exception can be thrown.

I'm removing them for now.  The only way you could test that this was
working would be to replace the mooseAssert in
MultiMooseEnum::operator[] with something that throws an exception...

Refs #6055.