Spherical or cylindrical gap geometry with higher dimensionality models ref #6161
Before this commit, the default behavior in MOOSE was to use the spherical
equations for 1d spherical models and cylindrical equations for 2d cylindrical
models. A higher-dimensionality model of a cylinder or sphere could only
use the flat plate equations for conductance.

This commit deprecates the 'coord_type' option that was previously used only
to override the default behavior on axisymmetric models.  It adds the following
new parameters:

gap_geometry_type = PLATE | CYLINDER | SPHERE
cylinder_axis_point_1
cylinder_axis_point_2
sphere_origin

With these options, the gap conductance model can now be specified independently
of the underlying coordinate system of the model, so these models can be used
for cylinders or spheres meshed in 2D (planar or axisymmetric) or 3D coordinate
systems. For those cases, the user must specify the points to define the geometry
of the cylinder or sphere.

This does not change any default behavior. It also consolidates previously
duplicated code.

Steve is working on an expanded set of tests for the new options enabled by this,
and will submit a pull request for that shortly after this is merged.