Do not divide by zero
When having both 1D mesh and 2D mesh in 3D space and when we use the
"projection" from the 2D side onto 1D mesh and the meshes are
aligned in z-direction, then the normal in PenetrationInfo will end
by with zero size. So when the normal is normalized, there is a
division by zero (caught by an assert statement in libMesh).

Closes #6179