Packed Range Fixup
The current code seems to run fine while receiving a iterator to a container
containing const Elem * and then subsequently using it without the constness.
The new packed range refactor however does not allow this. This explicit
const was added (and immediately cast away) to allow for the proper
type inference.

closes #6226