Add mutex in FunctionPeriodicBoundary::get_corresponding_pos().
This fixes a long standing issue with threaded execution of Periodic
BCs -- I don't believe they have ever worked correctly.  It turns out
that ParsedFunction evaluation is simply not thread-safe, despite the
fact that the FP_USE_THREAD_SAFE_EVAL macro in fparser.cc suggests
that you can make it so.

The best fix would be to actually fix FParser, but in the meantime I
think it would be really good to give users correct behavior, even
though it might degrade performance slightly.

Refs libMesh/libmesh#810.